Add ability to store token in local file
Constructors for fmDataAPI and fmAdminAPI now support the ['tokenFilePath'] option to specify a file path where the authentication token is read/written. Do NOT specify a file path that is web-visible. Added fmAPI::encryptToken and fmAPI::decryptToken so you can override and encrypt/decrypt the authentication token when it's stored in the $_SESSION  and/or token file.